requirements.txt
black==24.1.1
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.1
idna==3.6
jwt==1.3.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
pycparser==2.21
python-environ==0.4.54
requests==2.31.0
urllib3==2.1.0
